+++ packages/react-server-next/src/compat/server.tsx
@@ -1,7 +1,1 @@
-/** @todo */
-export class NextRequest extends Request {
-  nextUrl!: URL;
-}
-
-/** @todo */
-export class NextResponse extends Response {}
+export { NextRequest, NextResponse } from "@hiogawa/react-server/server";