+++ packages/react-server-next/src/vite/adapters/vercel-edge/entry.ts
@@ -1,0 +1,3 @@
+import { handler } from "@hiogawa/react-server/entry/ssr";
+
+export default handler;