+++ packages/react-server/examples/basic/e2e/basic.test.ts
@@ -821,6 +821,25 @@ test("server compnoent > fixture", async ({ page }) => {
   await page.getByText("TestDepServerComponent").click();
 });
 
+test("server-client-mixed package", async ({ page }) => {
+  await page.goto("/test/deps");
+  await page.getByText("TestDepMixed(Server)").click();
+  await waitForHydration(page);
+  await page.getByRole("button", { name: "TestDepMixed(Client): 0" }).click();
+  await page.getByRole("button", { name: "TestDepMixed(Client): 1" }).click();
+});
+
+test("tolerate export all with use client", async ({ page }) => {
+  await page.goto("/test/deps");
+  await waitForHydration(page);
+  await page
+    .getByRole("button", { name: "TestDepReExportExplicit: 0" })
+    .click();
+  await page
+    .getByRole("button", { name: "TestDepReExportExplicit: 1" })
+    .click();
+});
+
 test("client module used at boundary and non-boundary basic", async ({
   page,
 }) => {
@@ -895,7 +914,8 @@ testNoJs("redirect server component @nojs", async ({ page }) => {
   expect(await res.text()).toBe("");
 
   await page.goto("/test/redirect");
-  await page.getByRole("link", { name: "From Server Component" }).click();
+  await page.getByRole("link", { name: "Server Component" }).click();
+  await page.getByText("ok=server-component").click();
   await page.waitForURL("/test/redirect?ok=server-component");
 });
 
@@ -904,7 +924,8 @@ test("redirect server component @js", async ({ page }) => {
 
   await page.goto("/test/redirect");
   await waitForHydration(page);
-  await page.getByRole("link", { name: "From Server Component" }).click();
+  await page.getByRole("link", { name: "Server Component" }).click();
+  await page.getByText("ok=server-component").click();
   await page.waitForURL("/test/redirect?ok=server-component");
 });
 
@@ -912,15 +933,25 @@ testNoJs("redirect server action @nojs", async ({ page }) => {
   checkNoError(page);
 
   await page.goto("/test/redirect");
-  await page.getByRole("button", { name: "From Server Action" }).click();
+  await page.getByRole("button", { name: "Action" }).click();
+  await page.getByText("ok=server-action").click();
   await page.waitForURL("/test/redirect?ok=server-action");
 });
 
 test("redirect server action @js", async ({ page }) => {
   await page.goto("/test/redirect");
   await waitForHydration(page);
-  await page.getByRole("button", { name: "From Server Action" }).click();
+  await page.getByRole("button", { name: "Action" }).click();
   await page.waitForURL("/test/redirect?ok=server-action");
+});
+
+test("redirect suspense @js", async ({ page }) => {
+  await page.goto("/test/redirect");
+  await waitForHydration(page);
+  await page.getByRole("link", { name: "Suspense" }).click();
+  await page.getByText("fallback until redirect...").click();
+  await page.getByText("ok=suspense").click();
+  await page.waitForURL("/test/redirect?ok=suspense");
 });
 
 test("useActionState @js", async ({ page }) => {
@@ -1533,3 +1564,13 @@ async function testRouteGroups(page: Page) {
   await page.getByRole("heading", { name: "(marketing)/layout.tsx" }).click();
   await page.getByText("(marketing)/blog/page.tsx").click();
 }
+
+test("head inline script", async ({ page }) => {
+  checkNoError(page);
+  await page.goto("/test");
+  await waitForHydration(page);
+  const result = await page.evaluate(
+    () => (self as any).__testHeadInlineScript,
+  );
+  expect(result).toBe(true);
+});