+++ packages/react-server/src/features/server-action/redirect.ts
@@ -1,11 +1,12 @@
+import { tinyassert } from "@hiogawa/utils";
 import { isRedirectError } from "../error/shared";
 import type { RequestContext } from "../request-context/server";
 import type { RevalidationType } from "../server-component/utils";
 import type { ActionResult } from "./server";
 
-const ACTION_REDIRECT_KEY = "x-action-redirect";
+const FLIGHT_REDIRECT_KEY = "x-flight-redirect";
 
-type ActionRedirectMeta = {
+type FlightRedirectMeta = {
   location: string;
   revalidate?: RevalidationType;
 };
@@ -19,22 +20,18 @@ export function createActionRedirectResponse({
   isStream: boolean;
   requestContext: RequestContext;
 }): Response | undefined {
-  requestContext.revalidate;
   const error = actionResult.error;
   const redirect = error && isRedirectError(error);
   if (redirect) {
-    const headers = new Headers({
-      ...actionResult.responseHeaders,
-      ...error.headers,
-    });
+    const headers = new Headers(requestContext.getResponseHeaders());
     if (isStream) {
       headers.delete("location");
       headers.set(
-        ACTION_REDIRECT_KEY,
+        FLIGHT_REDIRECT_KEY,
         JSON.stringify({
           location: redirect.location,
           revalidate: requestContext.revalidate,
-        } satisfies ActionRedirectMeta),
+        } satisfies FlightRedirectMeta),
       );
       return new Response(null, {
         status: 200,
@@ -49,10 +46,31 @@ export function createActionRedirectResponse({
   return;
 }
 
-export function parseActionRedirectResponse(response: Response) {
-  const raw = response.headers.get(ACTION_REDIRECT_KEY);
+export function createFlightRedirectResponse(
+  response: Response,
+  requestContext: RequestContext,
+) {
+  const headers = new Headers(response.headers);
+  const location = headers.get("location");
+  tinyassert(typeof location === "string");
+  headers.delete("location");
+  headers.set(
+    FLIGHT_REDIRECT_KEY,
+    JSON.stringify({
+      location,
+      revalidate: requestContext.revalidate,
+    } satisfies FlightRedirectMeta),
+  );
+  return new Response(null, {
+    status: 200,
+    headers,
+  });
+}
+
+export function parseFlightRedirectResponse(response: Response) {
+  const raw = response.headers.get(FLIGHT_REDIRECT_KEY);
   if (raw) {
-    return JSON.parse(raw) as ActionRedirectMeta;
+    return JSON.parse(raw) as FlightRedirectMeta;
   }
   return;
 }