+++ packages/react-server/src/plugin/virtual.d.ts
@@ -1,0 +1,8 @@
+declare module "virtual:server-routes" {
+  const $default: Record<string, any>;
+  export default $default;
+
+  export const middleware:
+    | import("../features/next/middleware").MiddlewareModule
+    | undefined;
+}