+++ packages/react-server-next/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## v0.0.12 (2024-07-24)
+
+- fix: fix deployment with pre-rendered root path ([#587](https://github.com/hi-ogawa/vite-plugins/pull/587))
+
+## v0.0.11 (2024-07-22)
+
+- feat: support middleware ([#553](https://github.com/hi-ogawa/vite-plugins/pull/553))
+- chore: use plugin-react-swc ([#570](https://github.com/hi-ogawa/vite-plugins/pull/570))
+- chore: add `@vercel/og` example (node runtime) ([#558](https://github.com/hi-ogawa/vite-plugins/pull/558))
+
+## v0.0.10 (2024-07-14)
+
+- feat: vercel node runtime adapter ([#555](https://github.com/hi-ogawa/vite-plugins/pull/555))
+- fix: more compat exports (e.g. `next/server.js`, `Lusitana` font) ([#559](https://github.com/hi-ogawa/vite-plugins/pull/559))
+
 ## v0.0.9 (2024-07-11)
 
 - feat: support JSX in `.js` file ([#539](https://github.com/hi-ogawa/vite-plugins/pull/539))